requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.1.11
djangorestframework==3.9.2
idna==2.8
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
